requirements.txt
boto3==1.35.0
botocore==1.35.0
jsonschema==4.23.0
pytest==8.3.0
pytest-mock==3.14.0
python-docx==1.1.2